move adapt to gpu subdirectory
